fix(review): restore float32-before-resize in dct_energy_ratio, align create_sampler to float, black-format test_content_classifier
Round-2 review fixes:
- features.py _calculate_dct_energy_ratio: interpolate in float32 again
  (the typing pass had moved .astype after cv2.resize, quantising to uint8
  before interpolation -- a silent numeric change to an ML feature).
  Explicit np.ndarray annotation on first assignment keeps mypy clean.
- frame_sampler.py create_sampler: min_frames_threshold widened to float,
  matching FrameSampler.__init__ and all four strategy subclasses (the
  full-frames path passes float('inf')).
- test_content_classifier.py: black formatting only -- this was the sole
  black failure short-circuiting the Lint and Type Check job before the
  mypy step could run.

+++ src/giflab/sampling/frame_sampler.py
@@ -221,7 +221,7 @@ class FrameDifferenceCalculator:
 
 def create_sampler(
     strategy: SamplingStrategy,
-    min_frames_threshold: int = 30,
+    min_frames_threshold: float = 30,
     confidence_level: float = 0.95,
     verbose: bool = False,
     **kwargs: Any,
